requirements.txt
numpy==2.1.1
pandas==2.2.3
Pillow==10.4.0
streamlit==1.36.0
